64-bit LEB values are not always correctly decoded due to a casting issue, now they are.
<rdar://problem/27002247> 

llvm-svn: 274037

+++ lldb/source/Core/DataExtractor.cpp
@@ -1237,7 +1237,7 @@ DataExtractor::GetULEB128 (offset_t *offset_ptr) const
             while (src < end)
             {
                 uint8_t byte = *src++;
-                result |= (byte & 0x7f) << shift;
+                result |= (uint64_t)(byte & 0x7f) << shift;
                 if ((byte & 0x80) == 0)
                     break;
                 shift += 7;
@@ -1280,7 +1280,7 @@ DataExtractor::GetSLEB128 (offset_t *offset_ptr) const
         {
             bytecount++;
             byte = *src++;
-            result |= (byte & 0x7f) << shift;
+            result |= (int64_t)(byte & 0x7f) << shift;
             shift += 7;
             if ((byte & 0x80) == 0)
                 break;
